exercícios da aula do dia 9_SET do Curso de Python

+++ aula0/atv08.py
@@ -2,10 +2,10 @@
  calcule a sua área e a quantidade de tinta necessária para pintá-la, sabendo que cada litro de tinta
    pinta uma área de 2m². """
 
-base= int(input("Digite a largura da parede em metros: "))
-altura= int(input("Digite a altura da parede em metros: "))
-area= int(base*altura)
+base= float(input("Digite a largura da parede em metros: "))
+altura= float(input("Digite a altura da parede em metros: "))
+area= float(base*altura)
 
-tinta=int(area/2)
+tinta=float(area/2)
 
 print(f"A área total é {area}m² e serão necessários {tinta} litros de tinta para pintar essa parede!")
